+++ bin/inx-chronicle/src/stardust_inx/worker.rs
@@ -5,8 +5,8 @@ use std::fmt::Debug;
 
 use async_trait::async_trait;
 use chronicle::{
-    types::MessageId,
     runtime::{Actor, ActorContext, ActorError, Addr, HandleEvent, Report, SpawnActor},
+    types::MessageId,
 };
 use inx::{client::InxClient, proto::NoParams, tonic::Channel};
