+++ src/types/ledger/metadata.rs
@@ -1,0 +1,3 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
