+++ src/types/ledger/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod conflict_reason;
+mod inclusion_state;
