+++ src/types/message/mod.rs
@@ -1,0 +1,12 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod address;
+mod input;
+mod message;
+mod output;
+mod payload;
+mod signature;
+mod unlock_block;
+
+pub use self::{address::*, input::*, message::*, output::*, payload::*, signature::*, unlock_block::*};
