+++ src/types/message/output/foundry.rs
