+++ src/types/message/output/nft.rs
